requirements.txt
streamlit>=1.28.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
